changed use case registering, discord command registering  and new commands

+++ generic/models/substance.go
@@ -1,11 +1,11 @@
 package models
 
 type Substance struct {
-	ID                   uint   `json:"id"`
-	Label                string `json:"label"`
-	Value                string `json:"value"`
-	RecommendedDosageMin uint   `json:"recommended_dosage_min"`
-	RecommendedDosageMax uint   `json:"recommended_dosage_max"`
-	RecommendedPauseMin  uint   `json:"recommended_pause_min"`
-	RecommendedPauseMax  uint   `json:"recommended_pause_max"`
+	ID                   uint    `json:"id"`
+	Label                string  `json:"label"`
+	Value                string  `json:"value"`
+	RecommendedDosageMin float64 `json:"recommended_dosage_min"`
+	RecommendedDosageMax float64 `json:"recommended_dosage_max"`
+	RecommendedPauseMin  float64 `json:"recommended_pause_min"`
+	RecommendedPauseMax  float64 `json:"recommended_pause_max"`
 }
